1st Stage : Setting up
First of all, I chose a topic from the FlashCards. Flash cards can be used as an educational tool for a wide variety of tasks, from learning foreign words and scientific terms to remembering the birthdays of your new friends. I chose to create variety themed flashcards of JetBrains IDEs that can help choose which IDE to use for your next projects.